# website/static/openpanel-openapi.yaml
...
openapi: 3.0.3
info:
  title: OpenPanel API
  version: 2.0.2
  description: |-
    REST API for OpenPanel, a self-hosted web hosting control panel. Every feature available in the panel's web UI can also be accessed programmatically.

    All endpoints require a Bearer token obtained from `POST /api/login`. Beyond authentication, each endpoint also requires its documented feature to be enabled on the account's hosting plan.

    Success responses are JSON objects (`200`) unless noted otherwise (a handful of endpoints stream a file download instead). Error responses always include an `error` field and use standard HTTP status codes (400, 401, 403, 404, 409, 500, 503, 504).
  contact:
    name: OpenPanel
    url: https://openpanel.com
  license:
    name: OpenPanel EULA
    url: https://openpanel.com/LICENSE
servers:
- url: https://panel.example.com
  description: Your OpenPanel installation (default port 2083)
...
paths:
...
  /api/containers:
    get:
      summary: List compose services.
      operationId: get_containers
      tags:
      - Docker / Containers
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        default:
          $ref: '#/components/responses/Error'
      security:
      - bearerAuth: []
    post:
      summary: Add a new compose service.
      operationId: post_containers
      tags:
      - Docker / Containers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                service_name:
                  type: string
                  example: myservice
                image:
                  type: string
                  example: myimage:tag
                cpu:
                  type: string
                  example: '1'
                ram:
                  type: string
                  example: 512m
              example:
                service_name: myservice
                image: myimage:tag
                cpu: '1'
                ram: 512m
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        default:
          $ref: '#/components/responses/Error'
      security:
      - bearerAuth: []
  /api/containers/status:
    get:
      summary: List running container names.
      operationId: containers_status
      tags:
      - Docker / Containers
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        default:
          $ref: '#/components/responses/Error'
      security:
      - bearerAuth: []
  /api/containers/{service}/status:
    get:
      summary: Get container state and health.
      operationId: containers_service_status
      tags:
      - Docker / Containers
      parameters:
      - name: service
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        default:
          $ref: '#/components/responses/Error'
      security:
      - bearerAuth: []
  /api/containers/{service}/start:
    post:
      summary: Start container.
      operationId: containers_service_start
      tags:
      - Docker / Containers
      parameters:
      - name: service
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pull:
                  type: boolean
                  example: false
              example:
                pull: false
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        default:
          $ref: '#/components/responses/Error'
      security:
      - bearerAuth: []
  /api/containers/{service}/stop:
    post:
      summary: Stop container.
      operationId: containers_service_stop
      tags:
      - Docker / Containers
      parameters:
      - name: service
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        default:
          $ref: '#/components/responses/Error'
      security:
      - bearerAuth: []
  /api/containers/{service}/restart:
    post:
      summary: Restart container.
      operationId: containers_service_restart
      tags:
      - Docker / Containers
      parameters:
      - name: service
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        default:
          $ref: '#/components/responses/Error'
      security:
      - bearerAuth: []
  /api/containers/{service}/resources:
    patch:
      summary: Update CPU/RAM/PIDs limits.
      operationId: containers_service_resources
      tags:
      - Docker / Containers
      parameters:
      - name: service
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cpu:
                  type: string
                  example: '1'
                ram:
                  type: string
                  example: 512m
                pids:
                  type: string
                  example: '512'
              example:
                cpu: '1'
                ram: 512m
                pids: '512'
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        default:
          $ref: '#/components/responses/Error'
      security:
      - bearerAuth: []
  /api/containers/{service}/logs:
    get:
      summary: Tail container logs.
      operationId: containers_service_logs
      tags:
      - Docker / Containers
      parameters:
      - name: service
        in: path
        required: true
        schema:
          type: string
      - name: lines
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        default:
          $ref: '#/components/responses/Error'
      security:
      - bearerAuth: []
...
  /api/hosting/plan:
    get:
      summary: Get hosting plan details.
      operationId: hosting_plan
      tags:
      - Hosting Info
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        default:
          $ref: '#/components/responses/Error'
      security:
      - bearerAuth: []
...
  /api/services:
    get:
      summary: List all services with state.
      operationId: services
      tags:
      - System Services
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        default:
          $ref: '#/components/responses/Error'
      security:
      - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Obtain a token via `POST /api/login`. Tokens expire after 24 hours.
  responses:
    Error:
      description: An error occurred. See the `error` field for details.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: You do not own this domain
              hint:
                type: string
            required:
            - error
security:
- bearerAuth: []